[wpiutil] Rename Java package to edu.wpi.first.util
This is more consistent with wpimath being edu.wpi.first.math.
