+++ cscore/src/main/java/edu/wpi/first/cscore/CameraServerCvJNI.java
@@ -4,7 +4,7 @@
 
 package edu.wpi.first.cscore;
 
-import edu.wpi.first.wpiutil.RuntimeLoader;
+import edu.wpi.first.util.RuntimeLoader;
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.opencv.core.Core;
