+++ cscore/src/main/java/edu/wpi/first/cscore/CameraServerJNI.java
@@ -5,7 +5,7 @@
 package edu.wpi.first.cscore;
 
 import edu.wpi.first.cscore.raw.RawFrame;
-import edu.wpi.first.wpiutil.RuntimeLoader;
+import edu.wpi.first.util.RuntimeLoader;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.concurrent.atomic.AtomicBoolean;
