+++ hal/src/main/java/edu/wpi/first/hal/JNIWrapper.java
@@ -4,7 +4,7 @@
 
 package edu.wpi.first.hal;
 
-import edu.wpi.first.wpiutil.RuntimeLoader;
+import edu.wpi.first.util.RuntimeLoader;
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicBoolean;
