+++ wpilibj/src/dev/java/edu/wpi/first/wpilibj/DevMain.java
@@ -6,7 +6,7 @@ package edu.wpi.first.wpilibj;
 
 import edu.wpi.first.hal.HALUtil;
 import edu.wpi.first.networktables.NetworkTablesJNI;
-import edu.wpi.first.wpiutil.RuntimeDetector;
+import edu.wpi.first.util.RuntimeDetector;
 
 public final class DevMain {
   /** Main entry point. */
