+++ wpimath/src/main/java/edu/wpi/first/math/filter/MedianFilter.java
@@ -4,7 +4,7 @@
 
 package edu.wpi.first.math.filter;
 
-import edu.wpi.first.wpiutil.CircularBuffer;
+import edu.wpi.first.util.CircularBuffer;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
