+++ wpiutil/src/main/java/edu/wpi/first/util/ErrorMessages.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpiutil;
+package edu.wpi.first.util;
 
 import static java.util.Objects.requireNonNull;
