+++ wpiutil/src/main/java/edu/wpi/first/util/net/PortForwarder.java
@@ -2,9 +2,9 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpiutil.net;
+package edu.wpi.first.util.net;
 
-import edu.wpi.first.wpiutil.WPIUtilJNI;
+import edu.wpi.first.util.WPIUtilJNI;
 
 /**
  * Forward ports to another host. This is primarily useful for accessing Ethernet-connected devices
